requirements.txt
transformers==4.9.1
datasets
torch --extra-index-url https://download.pytorch.org/whl/cu113
wandb
tensorflow
packaging
gitpython
nltk
sentencepiece
matplotlib
pytorch-lightning==1.1.8
spacy 
sklearn
bert-score==0.3.9
tabulate
pandas
jsonlines
deepspeed==0.5.1
statsmodels


